% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDatasets.R
\docType{data}
\name{atp_2017}
\alias{atp_2017}
\title{atp_2017}
\format{
A data frame with 47 rows and 3 variables:
\describe{
\item{player_slug}{string. player names}
\item{ranking_points}{numeric. gained ATP points that year}
\item{wins}{numeric. ATP matches won that year}
}
}
\source{
https://github.com/datasets/atp-world-tour-tennis-data
}
\usage{
atp_2017
}
\description{
This datasets contains the name of the top male tennis players in 2017 as well as their tour wins and ranking points during that year.
}
\examples{
head(atp_2017)
}
\keyword{datasets}
