% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CubicSplineRegression.R
\name{dea_index}
\alias{dea_index}
\title{Find DEA Frontier Points}
\usage{
dea_index(xtab, ytab)
}
\arguments{
\item{xtab}{numeric vector of x-indices of data points}

\item{ytab}{numeric vector of y-indices of data points}
}
\value{
numeric vector with x-indices of points on the DEA-frontier
}
\description{
finds the x-indices of points on the DEA-frontier (with varying returns to scale)
}
\details{
The Data Envelopment Analysis (DEA) frontier under varying returns to scale (VRS) envelops all data points from above, forming a piecewise linear and monotonically increasing frontier.
This implementation follows the standard VRS approach as introduced by Banker, Charnes, and Cooper (1984).
The function returns the indices of all data points that lie on this frontier.
By linearly connecting these points, one obtains the VRS DEA frontier.
}
\examples{
xtab=atp_2017$ranking_points
ytab=atp_2017$wins
plot(xtab,ytab, xlab="ranking points", ylab="wins", main="ATP results in 2017 per player")
lines(xtab[dea_index(xtab,ytab)],
        ytab[dea_index(xtab,ytab)],
          type = "l", col = "red", lty = 2)
}
