% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CubicSplineRegression.R
\name{dea_lower_index}
\alias{dea_lower_index}
\title{Find DEA Lower Frontier Points}
\usage{
dea_lower_index(xtab, ytab)
}
\arguments{
\item{xtab}{numeric vector of x-indices of data points}

\item{ytab}{numeric vector of y-indices of data points}
}
\value{
numeric vector with x-indices of points on the lower DEA-frontier
The Data Envelopment Analysis (DEA) frontier under varying returns to scale (VRS) envelops all data points from below, forming a piecewise linear and monotonically increasing frontier.
This implementation follows the standard VRS approach as introduced by Banker, Charnes, and Cooper (1984).
The function returns the indices of all data points that lie on this frontier.
By linearly connecting these points, one obtains the lower VRS DEA frontier.
}
\description{
finds the x-indices of points on the lower DEA-frontier (with varying returns to scale)
}
\examples{
xtab=atp_2017$ranking_points
ytab=atp_2017$wins
plot(xtab,ytab, xlab="ranking points", ylab="wins", main="ATP results in 2017 per player")
lines(xtab[dea_lower_index(xtab,ytab)],
        ytab[dea_lower_index(xtab,ytab)],
          type = "l", col = "red", lty = 2)
}
