% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CubicSplineRegression.R
\name{fdh_lower_index}
\alias{fdh_lower_index}
\title{Find FDH Lower Frontier Points}
\usage{
fdh_lower_index(xtab, ytab)
}
\arguments{
\item{xtab}{numeric vector of x-indices of data points}

\item{ytab}{numeric vector of y-indices of data points}
}
\value{
numeric vector with x-indices of points on the lower FDH-frontier
}
\description{
finds the x-indices of points on the lower FDH-frontier
}
\details{
The free disposal hull method was developed in Deprins, Simar and Tulkens (1984). It is the lowest stair-case curve on or below all data points and by design monotonically increasing.
Formally, it is defined as \deqn{\phi(x)_n = \min \{ y_i, i:x_i \geq x \} }
The function returns the indices of all data points on the FDH (where \eqn{y_i = \phi(x_i)}).
By linearly connecting these points one arrives at the linearized FDH curve.
}
\examples{
xtab=atp_2017$ranking_points
ytab=atp_2017$wins
plot(xtab,ytab, xlab="ranking points", ylab="wins", main="ATP results in 2017 per player")
lines(xtab[fdh_lower_index(xtab,ytab)],
        ytab[fdh_lower_index(xtab,ytab)],
          type = "l", col = "red", lty = 2)
}
