% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDatasets.R
\docType{data}
\name{movies}
\alias{movies}
\title{movies}
\format{
A data frame with 60 rows and 5 variables:
\describe{
\item{Title}{string. movie title}
\item{Year}{numeric. year of release}
\item{Lifetime.Gross}{numeric. cumulated earnings over time in current US-dollars}
\item{Budget}{numeric. budget in current US-dollars}
\item{Runtime}{numeric. runtime in minutes}
}
}
\source{
https://www.kaggle.com/datasets/thedevastator/global-movie-franchise-revenue-and-budget-data
}
\usage{
movies
}
\description{
This datasets contains information about movies releases for big franchises since 2000.
}
\examples{
head(movies)
}
\keyword{datasets}
